Fixed visPoset
M2 now passes the labels of the elements and the relation matrix for the poset to the html template.  There are now methods to convert this data to the nodes and links arrays to use in the force layout.  We can now try to start adding poset editing and various poset tests from M2.