Poset editing issues fixed
I needed to empty out the references to nodes and links in the circle and path groups and create them all from scratch when nodes and links are rebuilt after editing the poset (except when just adding a new node).  This fixed the issues with the node names and the drag_line getting messed up when the poset was edited.