Updated visSimplicialComplex
Added methods to compute the facets of the current simplicial complex (needed for the M2 constructor) and to determine whether the current complex is pure.  Also wrote the M2 constructor, which has to construct a new polynomial ring if the user has changed the vertices, and modified how it was naming new nodes because M2 won't accept integers as variable names for polynomial rings.