Updated visSimplicialComplex
Added the boolean test "isPure," which just runs code in the javascript file itself to test the dimensions of the facets.  If we want to provide more boolean tests/invariants in the future, we will have to take into account the fact that we sometimes have to pass a new ring back to M2 in addition to the updated simplicial complex.  Also fixed a bug where the faces weren't getting un-highlighted when highlighting is disabled.