Update BootSideMenu to version 1.0.0 in Visualize package
Obtained from https://github.com/AndreaLombardo/BootSideMenu/tree/v1.0.0

This is the latest release of the library and is currently being packaged
for Debian.  Since the Debian Macaulay2 package will use Debian packages
of Javascript libraries instead of embedded ones, it would be nice to
keep the versions in sync.

Two small adjustments have been made for compatibility with this release:
* The `autoClose` option has been renamed `closeOnClick`.
* The default width of a side menu is now 15%, which is too narrow.  We
  now specify 230px, which was the previous default.