Update clipboard.js to version 2.6.0 in Visualize package
Obtained from https://github.com/zenorocha/clipboard.js/tree/v2.0.6

This is the latest release of the library and is the version currently
available in Debian.  Since the Debian Macaulay2 package will use
Debian packages of Javascript libraries instead of embedded ones, it
would be nice to keep the versions in sync.

One small adjustment has been made for compatibility with this release:
* `Clipboard` has been changed to `ClipboardJS`.