Update three.js to version r116 in Visualize package
Obtained from https://github.com/mrdoob/three.js/tree/r116

This is the latest release of the library and is (mostly) compatible
with r111, which is the version currently available in Debian.  Since
the Debian Macaulay2 package will use Debian packages of Javascript
libraries instead of embedded ones, it would be nice to keep the
versions in sync.

To make the transition, we've made several adjustments:
* `ShaderMaterial` no longer accepts `attributes`, so we no longer use
  it. Instead, we use `MeshBasicMaterial` for the cubes and
  `LineBasicMaterial` for their edges, which we define using
  `EdgesGeometry`.  After setting antialising to `true` when
  rendering, this provides a good approximation of the previous
  `ShaderMaterial` appearance.
* We need to explicitly set the background color to white, as the
  default is now black.