Create symlinks instead of copying Javascript files in Visualize
The copyJS function was failing to copy any files that happened to be
symbolic links, which was the case in the Debian/Ubuntu package
because many of these files were symbolic links to the packaged
versions of the corresponding Javascript libraries.

Instead, we create symbolic links to the directories containing these
files.  This also saves disk space since we won't be duplicating these
files for every call to "visualize".

We also ensure that the destination directory ends in a trailing slash
and create it if it doesn't already exist.